+++ SSD/file_handler_impl.h
@@ -30,7 +30,7 @@ public:
 	void rename(const string& old_name, const string& new_name) const override;
 
 	// Directory에서 file 이름의 len 만큼만 비교하여 같은 파일이 존재하는지 여부 반환
-	bool isFileExistByMatchLength(const string& dir_path, const string& file_name, int len) override;
+	bool isFileExistByMatchLength(const string& dir_path, const string& file_name, size_t len) override;
 
 	// Path에 있는 파일들 중에서 prefix가 포함되어 있는 파일명 반환
 	vector<string> getFileUsingPrefix(const string& path, const string& prefix) override;
